requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
alembic==1.13.3
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==42.0.8
dnspython==2.0.0
filetype==1.2.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
h11==0.16.0
idna==3.10
ipwhois==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.21.3
multidict==6.3.1
outcome==1.3.0.post0
packaging==24.2
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.5
pytz==2024.2
requests==2.32.3
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
validators==0.28.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
whois==1.20240129.2
wsproto==1.2.0
yarl==1.18.3
